feat(loader): add getConfigMetaFromFilename util
