+++ src/loader/impl/config.ts
@@ -1,17 +1,13 @@
 import * as path from 'path';
 import { Container } from '@artus/injection';
 import ConfigurationHandler from '../../configuration';
-import { ArtusInjectEnum, ARTUS_DEFAULT_CONFIG_ENV, CONFIG_PATTERN } from '../../constant';
+import { ArtusInjectEnum, CONFIG_PATTERN } from '../../constant';
 import { DefineLoader } from '../decorator';
 import { ManifestItem, Loader, LoaderFindOptions } from '../types';
 import compatibleRequire from '../../utils/compatible_require';
 import { isMatch } from '../../utils';
 import { Application } from '../../types';
-
-export interface ConfigFileMeta {
-  env: string;
-  namespace?: string;
-}
+import { getConfigMetaFromFilename } from '../utils/config_file_meta';
 
 @DefineLoader('config')
 class ConfigLoader implements Loader {
@@ -42,7 +38,7 @@ class ConfigLoader implements Loader {
   }
 
   async load(item: ManifestItem) {
-    const { namespace, env } = await this.getConfigFileMeta(item);
+    const { namespace, env } = getConfigMetaFromFilename(item.filename);
     let configObj = await this.loadConfigFile(item);
     if (namespace) {
       configObj = {
@@ -51,21 +47,6 @@ class ConfigLoader implements Loader {
     }
     this.configurationHandler.setConfig(env, configObj);
     return configObj;
-  }
-
-  protected async getConfigFileMeta(item: ManifestItem): Promise<ConfigFileMeta> {
-    let [namespace, env, extname] = item.filename.split('.');
-    if (!extname) {
-      // No env flag, set to Default
-      env = ARTUS_DEFAULT_CONFIG_ENV.DEFAULT;
-    }
-    const meta: ConfigFileMeta = {
-      env,
-    };
-    if (namespace !== 'config') {
-      meta.namespace = namespace;
-    }
-    return meta;
   }
 
   protected async loadConfigFile(item: ManifestItem): Promise<Record<string, any>> {
